Carga de datos completa

+++ Proyecto/src/Logica/Audio.java
@@ -7,7 +7,7 @@ public class Audio extends Publicacion implements Durable{
     private  boolean reproduciendo;
 
     //constructor
-    public Audio(int duracion, int velbits, int tiempoact, boolean reproduciendo, String n, int f, int cant, int e, int com) {
+    public Audio(int duracion, int velbits, int tiempoact, boolean reproduciendo, String n, long f, int cant, int e, int com) {
         super(n, f, cant, e, com);
         this.duracion = duracion;
         this.velbits = velbits;
